feat(model_selection.go): add OrchestratorResult struct to hold orchestration results for better organization and clarity refactor(orchestration_service.go): update orchestration methods to use the new OrchestratorResult struct for improved type safety and maintainability
